precision in the definition of few arrays is changed to satisfy compiler
Former-commit-id: 2c4382b35a11b71218bc8ff36ee4d045344a8a08

+++ PT/AMPS/src/pic/pic_batsrus_wrapper.f90
@@ -94,8 +94,8 @@ subroutine batsrus2amps_get_data_point(x,res,FoundFlag)
   real(8),dimension(0:nVar),intent(out)::res
   integer,intent(out)::FoundFlag
   
-  real,dimension(0:nVar)::State
-  real,dimension(nDim)::Xyz_D
+  real(8),dimension(0:nVar)::State
+  real(8),dimension(nDim)::Xyz_D
   logical::IsFound
   
   Xyz_D(:)=x(:)
